scripts/src/requirements.txt
PyYAML
google-cloud-bigquery
